Backend/requirements.txt
python-dotenv
pandas
numpy
setuptools
scikit-learn
flask
streamlit
fastapi
langchain_google_genai
pydantic
pinecone-client[grpc]
google-generativeai
vertexai
langchain-pinecone
langchain
uvicorn 
PyMuPDF
langchain_experimental
faiss-cpu
tiktoken
langchain-community 
-e .
